+++ controllers/helpers.js
@@ -74,7 +74,7 @@ function assignDeep(target, path, value, { mutate } = {}) {
 // ==============================
 //  STRING METHODS
 // ==============================
-exports._fn.string = { escapeChars, toKebabCase, readMongoId, readObjPath };
+exports._fn.string = { escapeChars, toKebabCase, readMongoId, readObjPath, inspectFileURL };
 
 /**
  * convert string to kebabCase
@@ -82,7 +82,7 @@ exports._fn.string = { escapeChars, toKebabCase, readMongoId, readObjPath };
  * @return {string}                         - kebabCased string
  */
 function toKebabCase(str) {
-    return str
+    return (str || '')
         .replace(/([a-z])([A-Z])/g, '$1-$2')                                // handle CamelCase
         .replace(/(-([A-Z])([A-Z]))/g, '-$2-$3')                            // handle signal words in CamelCase
         .replace(/[`'":;,.?!@#$%^&*_=~(){}<>/\\\[\]\-\+\|\s]+/g, '-')       // normalize special characters
@@ -126,6 +126,25 @@ function readObjPath(str) {
     return str.match(/[a-zA-Z0-9_$]+/g);
 }
 
+
+/**
+ * inspect the given file URL
+ * @param {string} str                      - any arbitrary string
+ * @param {array} extName                   - accepted file extension names
+ * @param {boolean} [raw=true]              - give the raw result(true) or string(false)
+ * @param {string} [use='https']            - set the use for the direct output
+ * @return {string|array|null}              - direct outputted string or raw result or null
+ */
+function inspectFileURL(str, extName, { raw = true , use = 'https' } = {}) {
+    const protocolExp   = '(?:(https?):\/\/)';
+    const fileNameExp   = `([0-9a-z\\s\\._%-]+\.(?:${extName.join('|')})`;
+    const domainExp     = '((?:[0-9a-z%-]+\\.(?!\\.))+[a-z]+)';
+    const pathExp       = '(\\/(?:[0-9a-z\\/\\s\\._%-]+\\/)?)';
+    const URLExp        = `^${protocolExp}?${domainExp}${pathExp}${fileNameExp}$)`;
+    const result        = new RegExp(URLExp, 'i').exec(str);
+    const filtrate      = Array.isArray(result) ? (use || result[1]) + '://' + result.slice(2, 5).join('') : null;
+    return raw === false ? filtrate : result;
+}
 
 
 // ==============================